HYD-4890 Add more tests to stub daddy to account for dependencies
HYD-4825 fixed an issue where the mock state was occasionally reported
incorrectly. Part of this change included the ability to specify
dependencies with a request only OR by specifying both a request and a
response. To cover this new ability I decided to add some more unit
tests.

This patch will simply update the mock-status and register-api tests
such that the tests cover the different ways in which dependencies are
specified and how they might be used.

Signed-off-by: Will Johnson <william.c.johnson@intel.com>