HYD-5096 Update Stub Daddy and Fix Race Condition
The realtime test is seeing a race condition in which we expect a call
to be made once but stub daddy is claiming that it is being called
twice. This is an intermittent error and does not always occur, but it
is much more prone to occur on CI. The purpose of this ticket is to
identity the cause of the race condition in stub daddy and fix it.

- Fix race condition
- Update stub daddy to use streams instead of promises
- Use eslint instead of jshint
- Replace wiretree with proxyquire
- Replace configulator with nconf
- Add more tests and update existing tests

Signed-off-by: Will Johnson <william.c.johnson@intel.com>