HYD-5320 Remove handle errors middleware from stub daddy and throw errors when necessary before route completes
After landing stub daddy 2.0 we realized there was an issue when running
realtime. The problem is that "handle errors" will throw an exception
after the route has written a response. This means that if you create a
mock the has a response of 500 it will cause stub daddy to crash, which
isn't the behavior we want. Instead, stub daddy should throw an
exception as soon as it knows there is a problem, before the route
completes. It should not throw errors based on the response statusCode.
This patch will address this issue.

Signed-off-by: Will Johnson <william.c.johnson@intel.com>