HYD-5485 Update stub daddy to use req@1.0.9 and remove headers from integration call options.
Now that req 1.0.9 has been published stub daddy should use this
version. The headers in the integration tests should also be removed as
they are forcing a `closed` connection.

This ticket will update the package.json file to use the new req version
and update both the integration tests and inline integration tests such
that the headers section is removed from the calls.

- Updated req version in package.json
- Removed call headers in inline-integration-test
- Removed call headers in integration-test

Signed-off-by: Will Johnson <william.c.johnson@intel.com>