HYD-5489 Add a clearTimeouts call to the afterTimeout middleware in stub daddy
Stub daddy has an afterTimeouts functionality that will hold a response
until a timeout has been reached. However, it is possible that the
server will be closed and the timeout will remain open, which could
potentially cause errors that are difficult to debug. To resolve this, a
clearTimeouts function should be called when the server is shut down.

This patch will add a clearTimeouts function to the afterTimeout
middleware and call the timeout from the close server function. Tests
will also be updated.

  - Added a clearTimeouts function to after-timeout
  - Updated after-timeout-test to test the clearTimeouts
  - Updated the server close function to clear out any timeouts when the
    server is closed in webservice
  - Updated webservice close test to ensure clearTimeouts is being
    called

Signed-off-by: Will Johnson <william.c.johnson@intel.com>